requirements.txt
yt-dlp
requests
Pillow
PyQt5
qtawesome
